Refactored LIS3DHTR dricer
removed LIS3DHTR_print_mmap()
moved LIS3DHTR_create_handle(), LIS3DHTR_read_reg() & LIS3DHTR_write_reg() to .c file
better naming for some macros

+++ code/CM7/Core/Src/freertos.c
@@ -163,7 +163,7 @@ void StartDefaultTask(void *argument)
 			float temp;
 			LIS3DHTR_get_temp( &LIS3DHTR_handle , &temp );
 
-			printf("Temp = %d\n" , temp );
+			printf("Temp = %d\n" , (int)temp );
 		}
 		if( input == 'a' )
 		{
